+++ src/Network/Connection.java
@@ -139,7 +139,7 @@ public class Connection {
 
      @throws IOException for any reason that ObjectOutputStream.writeObject() does
      */
-    public void sendCameraSpec(double zoom, int height, int width, boolean onOff, boolean manualAuto)
+    public void sendCameraSpec(int zoom, int height, int width, boolean onOff, boolean manualAuto)
             throws IOException {
         sendCameraSpec(new OutgoingCameraSpec(zoom, height, width, onOff, manualAuto));
     }
